+++ nail/src/pipeline/mod.rs
@@ -1,5 +1,245 @@
-mod align;
-pub use align::*;
+mod cloud_step;
+pub use cloud_step::*;
 
-mod search;
-pub use search::*;
+mod seed_step;
+pub use seed_step::*;
+
+mod align_step;
+pub use align_step::*;
+
+use std::collections::HashMap;
+use std::io::{stdout, Write};
+use std::path::PathBuf;
+use std::sync::{Arc, Mutex};
+
+use clap::Args;
+use rayon::iter::{IntoParallelRefIterator, IntoParallelRefMutIterator, ParallelIterator};
+use thiserror::Error;
+
+use crate::args::CommonArgs;
+use crate::util::PathBufExt;
+
+use libnail::align::structs::Alignment;
+use libnail::output::output_tabular::{Field, TableFormat};
+use libnail::structs::{Hmm, Profile, Sequence};
+
+#[derive(Error, Debug)]
+#[error("no profile with name: {profile_name}")]
+pub struct ProfileNotFoundError {
+    profile_name: String,
+}
+
+#[derive(Error, Debug)]
+#[error("no target with name: {target_name}")]
+pub struct TargetNotFoundError {
+    target_name: String,
+}
+
+#[derive(Args, Debug, Clone)]
+pub struct NailArgs {
+    /// Override the target database size (number of sequences) used for E-value calculation
+    #[arg(short = 'Z', value_name = "N")]
+    pub target_database_size: Option<usize>,
+    /// Pruning parameter alpha
+    #[arg(short = 'A', default_value_t = 12.0, value_name = "F")]
+    pub alpha: f32,
+    /// Pruning parameter beta
+    #[arg(short = 'B', default_value_t = 20.0, value_name = "F")]
+    pub beta: f32,
+    /// Pruning parameter gamma
+    #[arg(short = 'G', default_value_t = 5, value_name = "N")]
+    pub gamma: usize,
+    /// The P-value threshold for promoting hits past cloud search
+    #[arg(long = "cloud-thresh", default_value_t = 1e-3, value_name = "F")]
+    pub cloud_pvalue_threshold: f64,
+    /// The P-value threshold for promoting hits past forward
+    #[arg(long = "forward-thresh", default_value_t = 1e-4, value_name = "F")]
+    pub forward_pvalue_threshold: f64,
+    /// Compute the full dynamic programming matrices during alignment
+    #[arg(long, action)]
+    pub full_dp: bool,
+}
+
+#[derive(Args, Debug, Clone)]
+pub struct AlignOutputArgs {
+    /// Only report hits with an E-value below this value
+    #[arg(short = 'E', default_value_t = 10.0, value_name = "F")]
+    pub evalue_threshold: f64,
+    /// Where to place tabular output
+    #[arg(
+        short = 'T',
+        long = "tab-output",
+        default_value = "results.tsv",
+        value_name = "path"
+    )]
+    pub tsv_results_path: PathBuf,
+    /// Where to place alignment output
+    #[arg(short = 'O', long = "output", value_name = "path")]
+    pub ali_results_path: Option<PathBuf>,
+}
+
+#[derive(Debug, Args)]
+pub struct AlignArgs {
+    /// Query file
+    #[arg(value_name = "QUERY.[fasta:hmm:sto]")]
+    pub query_path: PathBuf,
+    /// Target file
+    #[arg(value_name = "TARGET.fasta")]
+    pub target_path: PathBuf,
+    /// Alignment seeds from running nail seed (or elsewhere)
+    #[arg(value_name = "SEEDS.json")]
+    pub seeds_path: PathBuf,
+
+    /// Arguments that are passed to nail functions
+    #[command(flatten)]
+    pub nail_args: NailArgs,
+
+    /// Arguments that control output options
+    #[command(flatten)]
+    pub output_args: AlignOutputArgs,
+
+    /// Arguments that are common across all nail subcommands
+    #[command(flatten)]
+    pub common_args: CommonArgs,
+}
+
+pub const DEFAULT_COLUMNS: [Field; 10] = [
+    Field::Target,
+    Field::Query,
+    Field::TargetStart,
+    Field::TargetEnd,
+    Field::QueryStart,
+    Field::QueryEnd,
+    Field::Score,
+    Field::CompBias,
+    Field::Evalue,
+    Field::CellFrac,
+];
+
+#[derive(Clone)]
+pub struct Pipeline {
+    pub seed: Box<dyn SeedStep + Send + Sync>,
+    pub cloud_search: Box<dyn CloudSearchStep + Send + Sync>,
+    pub align: Box<dyn AlignStep + Send + Sync>,
+}
+
+impl Pipeline {
+    fn run(
+        &mut self,
+        profile: &mut Profile,
+        targets: &HashMap<String, Sequence>,
+    ) -> Option<Vec<Alignment>> {
+        self.seed.run(profile, targets).map(|seeds| {
+            seeds
+                .iter()
+                .filter_map(|(target_name, seed)| {
+                    let target = targets.get(target_name).unwrap();
+                    self.cloud_search
+                        .run(profile, target, seed)
+                        .and_then(|bounds| self.align.run(profile, target, bounds))
+                })
+                .collect()
+        })
+    }
+}
+
+pub enum HeaderStatus {
+    Unwritten,
+    Written,
+}
+
+pub struct Output {
+    alignment_writer: Box<dyn Write + Send>,
+    table_writer: Box<dyn Write + Send>,
+    table_format: TableFormat,
+    header_status: HeaderStatus,
+}
+
+impl Output {
+    pub fn new(args: &AlignOutputArgs) -> anyhow::Result<Self> {
+        Ok(Self {
+            alignment_writer: match &args.ali_results_path {
+                Some(path) => Box::new(path.open(true)?),
+                None => Box::new(stdout()),
+            },
+            table_writer: Box::new(args.tsv_results_path.open(true)?),
+            table_format: TableFormat::new(&DEFAULT_COLUMNS)?,
+            header_status: HeaderStatus::Unwritten,
+        })
+    }
+
+    pub fn write(&mut self, alignments: &mut [Alignment]) -> anyhow::Result<()> {
+        self.table_format.reset_widths();
+        self.table_format.update_widths(alignments);
+
+        alignments.sort_by(|a, b| a.e_value.partial_cmp(&b.e_value).unwrap());
+
+        if let HeaderStatus::Unwritten = self.header_status {
+            let header = TableFormat::header(&self.table_format)?;
+            writeln!(self.table_writer, "{header}")?;
+            self.header_status = HeaderStatus::Written;
+        }
+
+        alignments.iter().for_each(|ali| {
+            writeln!(
+                self.table_writer,
+                "{}",
+                ali.tab_string_formatted(&self.table_format)
+            )
+            .expect("failed to write tabular output");
+
+            writeln!(self.alignment_writer, "{}", ali.ali_string())
+                .expect("failed to write alignment output");
+        });
+
+        Ok(())
+    }
+}
+
+pub fn run_pipeline_profile_to_sequence(
+    queries: &mut [Profile],
+    targets: &HashMap<String, Sequence>,
+    pipeline: Pipeline,
+    output: Arc<Mutex<Output>>,
+) {
+    queries.par_iter_mut().for_each_with(
+        (pipeline, targets, output),
+        |(pipeline, targets, output), profile| {
+            let mut alignments: Vec<Alignment> = pipeline.run(profile, targets).unwrap();
+
+            match output.lock() {
+                Ok(mut guard) => {
+                    guard.write(&mut alignments).unwrap();
+                }
+                Err(_) => panic!(),
+            }
+        },
+    )
+}
+
+pub fn run_pipeline_sequence_to_sequence(
+    queries: &[Sequence],
+    targets: &HashMap<String, Sequence>,
+    pipeline: Pipeline,
+    output: Arc<Mutex<Output>>,
+) {
+    queries.par_iter().for_each_with(
+        (pipeline, targets, output),
+        |(pipeline, targets, output), sequence| {
+            let mut profile = Hmm::from_blosum_62_and_sequence(sequence)
+                .map(|h| Profile::new(&h))
+                .expect("failed to build profile from sequence");
+
+            profile.calibrate_tau(200, 100, 0.04);
+
+            let mut alignments: Vec<Alignment> = pipeline.run(&mut profile, targets).unwrap();
+
+            match output.lock() {
+                Ok(mut guard) => {
+                    guard.write(&mut alignments).unwrap();
+                }
+                Err(_) => panic!(),
+            }
+        },
+    )
+}
